+++ parent/http/src/main/java/com/dcare/ao/AddBloodGlucoseAO.java
@@ -10,7 +10,7 @@ package com.dcare.ao;
 public class AddBloodGlucoseAO {
 	private int familyUserId;  //家庭用户id
 	
-	private int glucose; //血糖值
+	private float glucose; //血糖值
 	
 	private int meal; //哪一餐
 	
@@ -27,11 +27,11 @@ public class AddBloodGlucoseAO {
 		this.familyUserId = familyUserId;
 	}
 
-	public int getGlucose() {
+	public float getGlucose() {
 		return glucose;
 	}
 
-	public void setGlucose(int glucose) {
+	public void setGlucose(float glucose) {
 		this.glucose = glucose;
 	}
